@@ -9,7 +9,7 @@ authors = ["dignifiedquire <me@dignifiedquire.com>", "n0 team"]
 repository = "https://github.com/n0-computer/iroh"
 
 # Sadly this also needs to be updated in .github/workflows/ci.yml
-rust-version = "1.71"
+rust-version = "1.72"
 
 [dependencies]
 aead = { version = "0.5.2", features = ["bytes"] }
@@ -100,6 +100,7 @@ ntest = "0.9"
 pretty_assertions = "1.4"
 proptest = "1.2.0"
 rand_chacha = "0.3.1"
+testdir = "0.8"
 tokio = { version = "1", features = ["io-util", "sync", "rt", "net", "fs", "macros", "time", "test-util"] }
 tracing-subscriber = { version = "0.3", features = ["env-filter"] }
 iroh-test = { path = "../iroh-test" }
