@@ -1,19 +1,23 @@
 use std::{
-    collections::{hash_map, HashMap, HashSet},
-    net::IpAddr,
+    collections::{HashMap, HashSet},
+    net::{IpAddr, SocketAddr},
     sync::{atomic::Ordering, Arc},
     time::{Duration, Instant},
 };
 
+use anyhow::Context;
 use backoff::backoff::Backoff;
 use bytes::{Bytes, BytesMut};
 use iroh_metrics::{inc, inc_by};
-use tokio::{sync::mpsc, time};
-use tokio_util::sync::CancellationToken;
-use tracing::{debug, info, trace, warn};
+use tokio::{
+    sync::{mpsc, oneshot},
+    task::{JoinHandle, JoinSet},
+    time,
+};
+use tracing::{debug, info, info_span, trace, warn, Instrument};
 
 use crate::{
-    derp::{self, MAX_PACKET_SIZE},
+    derp::{self, http::ClientError, ReceivedMessage, MAX_PACKET_SIZE},
     key::{PublicKey, PUBLIC_KEY_LENGTH},
 };
 
@@ -36,10 +40,6 @@ pub(super) enum DerpActorMessage {
         region_id: u16,
         peer: Option<PublicKey>,
     },
-    CloseOrReconnect {
-        region_id: u16,
-        reason: &'static str,
-    },
     NotePreferred(u16),
     MaybeCloseDerpsOnRebind(Vec<IpAddr>),
     Shutdown,
@@ -48,35 +48,228 @@ pub(super) enum DerpActorMessage {
 /// Contains fields for an active DERP connection.
 #[derive(Debug)]
 struct ActiveDerp {
-    c: derp::http::Client,
-    cancel: CancellationToken,
     /// The time of the last request for its write
     /// channel (currently even if there was no write).
     last_write: Instant,
-    create_time: Instant,
-    reader: ReaderState,
+    msg_sender: mpsc::Sender<ActorMessage>,
+    /// Contains optional alternate routes to use as an optimization instead of
+    /// contacting a peer via their home DERP connection. If they sent us a message
+    /// on this DERP connection (which should really only be on our DERP
+    /// home connection, or what was once our home), then we remember that route here to optimistically
+    /// use instead of creating a new DERP connection back to their home.
+    derp_routes: Vec<PublicKey>,
+    region: u16,
+    derp_client: derp::http::Client,
+    derp_client_receiver: derp::http::ClientReceiver,
+    /// The set of senders we know are present on this connection, based on
+    /// messages we've received from the server.
+    peer_present: HashSet<PublicKey>,
+    backoff: backoff::exponential::ExponentialBackoff<backoff::SystemClock>,
+    last_packet_time: Option<Instant>,
+    last_packet_src: Option<PublicKey>,
 }
 
-/// A route entry for a public key, saying that a certain peer should be available at DERP
-/// node derpID, as long as the current connection for that derpID is dc. (but dc should not be
-/// used to write directly; it's owned by the read/write loops)
 #[derive(Debug)]
-struct DerpRoute {
-    derp_id: u16,
-    dc: derp::http::Client, // don't use directly; see comment above
+#[allow(clippy::large_enum_variant)]
+enum ActiveDerpMessage {
+    GetLastWrite(oneshot::Sender<Instant>),
+    Ping(oneshot::Sender<Result<Duration, ClientError>>),
+    GetLocalAddr(oneshot::Sender<Option<SocketAddr>>),
+    GetPeerRoute(PublicKey, oneshot::Sender<Option<derp::http::Client>>),
+    GetClient(oneshot::Sender<derp::http::Client>),
+    NotePreferred(bool),
+    Shutdown,
+}
+
+impl ActiveDerp {
+    fn new(
+        region: u16,
+        derp_client: derp::http::Client,
+        derp_client_receiver: derp::http::ClientReceiver,
+        msg_sender: mpsc::Sender<ActorMessage>,
+    ) -> Self {
+        ActiveDerp {
+            last_write: Instant::now(),
+            msg_sender,
+            derp_routes: Default::default(),
+            region,
+            peer_present: HashSet::new(),
+            backoff: backoff::exponential::ExponentialBackoffBuilder::new()
+                .with_initial_interval(Duration::from_millis(10))
+                .with_max_interval(Duration::from_secs(5))
+                .build(),
+            last_packet_time: None,
+            last_packet_src: None,
+            derp_client,
+            derp_client_receiver,
+        }
+    }
+
+    async fn run(mut self, mut inbox: mpsc::Receiver<ActiveDerpMessage>) -> anyhow::Result<()> {
+        self.derp_client
+            .connect()
+            .await
+            .context("initial connection")?;
+
+        loop {
+            tokio::select! {
+                Some(msg) = inbox.recv() => {
+                    match msg {
+                        ActiveDerpMessage::GetLastWrite(r) => {
+                            r.send(self.last_write).ok();
+                        }
+                        ActiveDerpMessage::Ping(r) => {
+                            r.send(self.derp_client.ping().await).ok();
+                        }
+                        ActiveDerpMessage::GetLocalAddr(r) => {
+                            r.send(self.derp_client.local_addr().await).ok();
+                        }
+                        ActiveDerpMessage::GetClient(r) => {
+                            self.last_write = Instant::now();
+                            r.send(self.derp_client.clone()).ok();
+                        }
+                        ActiveDerpMessage::NotePreferred(is_preferred) => {
+                            self.derp_client.note_preferred(is_preferred).await;
+                        }
+                        ActiveDerpMessage::GetPeerRoute(peer, r) => {
+                            let res = if self.derp_routes.contains(&peer) {
+                                Some(self.derp_client.clone())
+                            } else {
+                                None
+                            };
+                            r.send(res).ok();
+                        }
+                        ActiveDerpMessage::Shutdown => {
+                            self.derp_client.close().await.ok();
+                            break;
+                        }
+                    }
+                }
+                msg = self.derp_client_receiver.recv() => {
+                    if let Some(msg) = msg {
+                        if self.handle_derp_msg(msg).await == ReadResult::Break {
+                            // fatal error
+                            self.derp_client.close().await.ok();
+                            break;
+                        }
+                    }
+                }
+                else => {
+                    break;
+                }
+            }
+        }
+
+        Ok(())
+    }
+
+    async fn handle_derp_msg(
+        &mut self,
+        msg: Result<(ReceivedMessage, usize), ClientError>,
+    ) -> ReadResult {
+        match msg {
+            Err(err) => {
+                warn!("recv error {:?}", err);
+
+                // Forget that all these peers have routes.
+                let peers: Vec<_> = self.peer_present.drain().collect();
+                self.derp_routes.retain(|peer| !peers.contains(peer));
+
+                if matches!(
+                    err,
+                    derp::http::ClientError::Closed | derp::http::ClientError::IPDisabled
+                ) {
+                    // drop client
+                    return ReadResult::Break;
+                }
+
+                // If our DERP connection broke, it might be because our network
+                // conditions changed. Start that check.
+                // TODO:
+                // self.re_stun("derp-recv-error").await;
+
+                // Back off a bit before reconnecting.
+                match self.backoff.next_backoff() {
+                    Some(t) => {
+                        debug!("backoff sleep: {}ms", t.as_millis());
+                        time::sleep(t).await;
+                        ReadResult::Continue
+                    }
+                    None => ReadResult::Break,
+                }
+            }
+            Ok((msg, conn_gen)) => {
+                // reset
+                self.backoff.reset();
+                let now = Instant::now();
+                if self.last_packet_time.is_none()
+                    || self.last_packet_time.as_ref().unwrap().elapsed() > Duration::from_secs(5)
+                {
+                    self.last_packet_time = Some(now);
+                }
+
+                match msg {
+                    derp::ReceivedMessage::ServerInfo { .. } => {
+                        info!(%conn_gen, "connected");
+                        ReadResult::Continue
+                    }
+                    derp::ReceivedMessage::ReceivedPacket { source, data } => {
+                        trace!(len=%data.len(), "received msg");
+                        // If this is a new sender we hadn't seen before, remember it and
+                        // register a route for this peer.
+                        if self.last_packet_src.is_none()
+                            || &source != self.last_packet_src.as_ref().unwrap()
+                        {
+                            // avoid map lookup w/ high throughput single peer
+                            self.last_packet_src = Some(source);
+                            if !self.peer_present.contains(&source) {
+                                self.peer_present.insert(source);
+                                self.derp_routes.push(source);
+                            }
+                        }
+
+                        let res = DerpReadResult {
+                            region_id: self.region,
+                            src: source,
+                            buf: data,
+                        };
+                        self.msg_sender
+                            .send(ActorMessage::ReceiveDerp(res))
+                            .await
+                            .ok();
+                        ReadResult::Continue
+                    }
+                    derp::ReceivedMessage::Ping(data) => {
+                        // Best effort reply to the ping.
+                        let dc = self.derp_client.clone();
+                        tokio::task::spawn(async move {
+                            if let Err(err) = dc.send_pong(data).await {
+                                warn!("pong error: {:?}", err);
+                            }
+                        });
+                        ReadResult::Continue
+                    }
+                    derp::ReceivedMessage::Health { .. } => ReadResult::Continue,
+                    derp::ReceivedMessage::PeerGone(key) => {
+                        self.derp_routes.retain(|peer| peer != &key);
+                        ReadResult::Continue
+                    }
+                    _ => {
+                        // Ignore.
+                        ReadResult::Continue
+                    }
+                }
+            }
+        }
+    }
 }
 
 pub(super) struct DerpActor {
     conn: Arc<Inner>,
     /// DERP regionID -> connection to a node in that region
-    active_derp: HashMap<u16, ActiveDerp>,
-    /// Contains optional alternate routes to use as an optimization instead of
-    /// contacting a peer via their home DERP connection.  If they sent us a message
-    /// on a different DERP connection (which should really only be on our DERP
-    /// home connection, or what was once our home), then we remember that route here to optimistically
-    /// use instead of creating a new DERP connection back to their home.
-    derp_route: HashMap<PublicKey, DerpRoute>,
+    active_derp: HashMap<u16, (mpsc::Sender<ActiveDerpMessage>, JoinHandle<()>)>,
     msg_sender: mpsc::Sender<ActorMessage>,
+    ping_tasks: JoinSet<(u16, bool)>,
 }
 
 impl DerpActor {
@@ -84,8 +277,8 @@ impl DerpActor {
         DerpActor {
             conn,
             active_derp: HashMap::default(),
-            derp_route: HashMap::new(),
             msg_sender,
+            ping_tasks: Default::default(),
         }
     }
 
@@ -97,6 +290,12 @@ impl DerpActor {
 
         loop {
             tokio::select! {
+                Some(Ok((region_id, ping_success))) = self.ping_tasks.join_next() => {
+                    if !ping_success {
+                        self.close_or_reconnect_derp(region_id, "rebind-ping-fail")
+                            .await;
+                    }
+                }
                 Some(msg) = receiver.recv() => {
                     match msg {
                         DerpActorMessage::Send { region_id, contents, peer } => {
@@ -104,9 +303,6 @@ impl DerpActor {
                         }
                         DerpActorMessage::Connect { region_id, peer } => {
                             self.connect_derp(region_id, peer.as_ref()).await;
-                        }
-                        DerpActorMessage::CloseOrReconnect { region_id, reason } => {
-                            self.close_or_reconnect_derp(region_id, reason).await;
                         }
                         DerpActorMessage::NotePreferred(my_derp) => {
                             self.note_preferred(my_derp).await;
@@ -118,28 +314,6 @@ impl DerpActor {
                             debug!("shutting down");
                             self.close_all_derp("conn-close").await;
                             break;
-                        }
-                    }
-                }
-                (region_id, result, action) = self.recv_all() => {
-                    trace!("tick: recvs: {:?}, {:?}", result, action);
-                    match action {
-                        ReadAction::None => {},
-                        ReadAction::AddPeerRoutes { peers, region, derp_client } => {
-                            self.add_derp_peer_routes(peers, region, derp_client);
-                        },
-                        ReadAction::RemovePeerRoutes { peers, region, derp_client } => {
-                            self.remove_derp_peer_routes(peers, region, &derp_client);
-                        }
-                    }
-                    match result {
-                        ReadResult::Break => {
-                            // drop client
-                            self.close_derp(region_id, "read error").await;
-                        }
-                        ReadResult::Continue => {}
-                        ReadResult::Yield(read_result) => {
-                            self.msg_sender.send(ActorMessage::ReceiveDerp(read_result)).await.ok();
                         }
                     }
                 }
@@ -155,28 +329,17 @@ impl DerpActor {
         }
     }
 
-    async fn recv_all(&mut self) -> (u16, ReadResult, ReadAction) {
-        if self.active_derp.is_empty() {
-            futures::future::pending::<(u16, ReadResult, ReadAction)>().await;
-        }
-
-        let ((region, (result, action)), _, _) =
-            futures::future::select_all(self.active_derp.iter_mut().map(|(region, ad)| {
-                Box::pin(async move {
-                    let res = ad.reader.recv().await;
-                    (*region, res)
-                })
-            }))
-            .await;
-
-        (region, result, action)
-    }
-
     async fn note_preferred(&self, my_num: u16) {
-        futures::future::join_all(self.active_derp.iter().map(|(i, ad)| async move {
-            let b = *i == my_num;
-            ad.c.note_preferred(b).await;
-        }))
+        futures::future::join_all(
+            self.active_derp
+                .iter()
+                .map(|(region_id, (s, _))| async move {
+                    let is_preferred = *region_id == my_num;
+                    s.send(ActiveDerpMessage::NotePreferred(is_preferred))
+                        .await
+                        .ok()
+                }),
+        )
         .await;
     }
 
@@ -190,7 +353,7 @@ impl DerpActor {
         // Derp Send
         let derp_client = self.connect_derp(region_id, Some(&peer)).await;
         for content in &contents {
-            trace!("[DERP] -> {} ({}b) {:?}", region_id, content.len(), peer);
+            trace!(region_id, ?peer, "sending {}B", content.len());
         }
         let total_bytes = contents.iter().map(|c| c.len() as u64).sum::<u64>();
 
@@ -206,7 +369,7 @@ impl DerpActor {
                     inc_by!(MagicsockMetrics, send_derp, total_bytes);
                 }
                 Err(err) => {
-                    warn!("derp.send: failed {:?}", err);
+                    warn!(region_id, "send: failed {:?}", err);
                     inc!(MagicsockMetrics, send_derp_error);
                 }
             }
@@ -216,6 +379,20 @@ impl DerpActor {
         let mut wakers = self.conn.network_send_wakers.lock().unwrap();
         if let Some(waker) = wakers.take() {
             waker.wake();
+        }
+    }
+
+    /// Returns `true`if the message was sent successfully.
+    async fn send_to_active(&mut self, region: u16, msg: ActiveDerpMessage) -> bool {
+        match self.active_derp.get(&region) {
+            Some((s, _)) => match s.send(msg).await {
+                Ok(_) => true,
+                Err(mpsc::error::SendError(_)) => {
+                    self.close_derp(region, "sender-closed").await;
+                    false
+                }
+            },
+            None => false,
         }
     }
 
@@ -229,9 +406,16 @@ impl DerpActor {
         // well use it. (It's a little arbitrary whether we use this one vs. the reverse route
         // below when we have both.)
 
-        if let Some(ad) = self.active_derp.get_mut(&region_id) {
-            ad.last_write = Instant::now();
-            return ad.c.clone();
+        {
+            let (os, or) = oneshot::channel();
+            if self
+                .send_to_active(region_id, ActiveDerpMessage::GetClient(os))
+                .await
+            {
+                if let Ok(client) = or.await {
+                    return client;
+                }
+            }
         }
 
         // If we don't have an open connection to the peer's home DERP
@@ -241,11 +425,20 @@ impl DerpActor {
         // node in SF to reach us, so we can reply to them using our
         // SF connection rather than dialing Frankfurt.
         if let Some(peer) = peer {
-            if let Some(r) = self.derp_route.get(peer) {
-                if let Some(ad) = self.active_derp.get_mut(&r.derp_id) {
-                    if ad.c == r.dc {
-                        ad.last_write = Instant::now();
-                        return ad.c.clone();
+            for region_id in self
+                .active_derp
+                .keys()
+                .copied()
+                .collect::<Vec<_>>()
+                .into_iter()
+            {
+                let (os, or) = oneshot::channel();
+                if self
+                    .send_to_active(region_id, ActiveDerpMessage::GetPeerRoute(*peer, os))
+                    .await
+                {
+                    if let Ok(Some(client)) = or.await {
+                        return client;
                     }
                 }
             }
@@ -263,7 +456,7 @@ impl DerpActor {
         let ipv6_reported = self.conn.ipv6_reported.clone();
 
         // building a client does not dial
-        let dc = derp::http::ClientBuilder::new()
+        let (dc, dc_receiver) = derp::http::ClientBuilder::new()
             .address_family_selector(move || {
                 let ipv6_reported = ipv6_reported.clone();
                 Box::pin(async move { ipv6_reported.load(Ordering::Relaxed) })
@@ -283,27 +476,23 @@ impl DerpActor {
             .build(self.conn.secret_key.clone())
             .expect("will only fail is a `get_region` callback is not supplied");
 
-        let cancel = CancellationToken::new();
-        let ad = ActiveDerp {
-            c: dc.clone(),
-            cancel: cancel.clone(),
-            last_write: Instant::now(),
-            create_time: Instant::now(),
-            reader: ReaderState::new(region_id, cancel, dc.clone()),
-        };
+        let (s, r) = mpsc::channel(64);
+
+        let c = dc.clone();
+        let msg_sender = self.msg_sender.clone();
+        let handle = tokio::task::spawn(
+            async move {
+                let ad = ActiveDerp::new(region_id, c, dc_receiver, msg_sender);
+
+                if let Err(err) = ad.run(r).await {
+                    warn!("connection error: {:?}", err);
+                }
+            }
+            .instrument(info_span!("active-derp", %region_id)),
+        );
 
         // Insert, to make sure we do not attempt to double connect.
-        self.active_derp.insert(region_id, ad);
-
-        // Kickoff a connection establishment in the background
-        let dc_spawn = dc.clone();
-        tokio::task::spawn(async move {
-            // Make sure we can establish a connection.
-            if let Err(err) = dc_spawn.connect().await {
-                // TODO: what to do?
-                warn!("failed to connect to derp server: {:?}", err);
-            }
-        });
+        self.active_derp.insert(region_id, (s, handle));
 
         inc!(MagicsockMetrics, num_derp_conns_added);
 
@@ -321,37 +510,55 @@ impl DerpActor {
     /// and pings all those that do.
     async fn maybe_close_derps_on_rebind(&mut self, okay_local_ips: &[IpAddr]) {
         let mut tasks = Vec::new();
-        for (region_id, ad) in &self.active_derp {
-            let la = match ad.c.local_addr().await {
-                None => {
-                    tasks.push((*region_id, "rebind-no-localaddr"));
-                    continue;
+        for region_id in self
+            .active_derp
+            .keys()
+            .copied()
+            .collect::<Vec<_>>()
+            .into_iter()
+        {
+            let (os, or) = oneshot::channel();
+            let la = if self
+                .send_to_active(region_id, ActiveDerpMessage::GetLocalAddr(os))
+                .await
+            {
+                match or.await {
+                    Ok(None) | Err(_) => {
+                        tasks.push((region_id, "rebind-no-localaddr"));
+                        continue;
+                    }
+                    Ok(Some(la)) => la,
                 }
-                Some(la) => la,
+            } else {
+                tasks.push((region_id, "rebind-no-localaddr"));
+                continue;
             };
 
             if !okay_local_ips.contains(&la.ip()) {
-                tasks.push((*region_id, "rebind-default-route-change"));
+                tasks.push((region_id, "rebind-default-route-change"));
                 continue;
             }
 
-            let dc = ad.c.clone();
-            let region_id = *region_id;
-            let msg_sender = self.msg_sender.clone();
-            tokio::task::spawn(time::timeout(Duration::from_secs(3), async move {
-                if let Err(_err) = dc.ping().await {
-                    msg_sender
-                        .send(ActorMessage::CloseOrReconnect(
-                            region_id,
-                            "rebind-ping-fail",
-                        ))
-                        .await
-                        .unwrap();
-                    return;
-                }
-                debug!("post-rebind ping of DERP region {} okay", region_id);
-            }));
+            let (os, or) = oneshot::channel();
+            let ping_sent = self
+                .send_to_active(region_id, ActiveDerpMessage::Ping(os))
+                .await;
+
+            self.ping_tasks.spawn(async move {
+                let ping_success = time::timeout(Duration::from_secs(3), async {
+                    if ping_sent {
+                        or.await.is_ok()
+                    } else {
+                        false
+                    }
+                })
+                .await
+                .unwrap_or(false);
+
+                (region_id, ping_success)
+            });
         }
+
         for (region_id, why) in tasks {
             self.close_or_reconnect_derp(region_id, why).await;
         }
@@ -369,22 +576,35 @@ impl DerpActor {
     }
 
     async fn clean_stale_derp(&mut self) {
-        debug!("cleanup {} derps", self.active_derp.len());
+        trace!("checking {} derps for staleness", self.active_derp.len());
         let now = Instant::now();
 
         let mut to_close = Vec::new();
-        for (i, ad) in &self.active_derp {
+        for (i, (s, _)) in &self.active_derp {
             if *i == self.conn.my_derp() {
                 continue;
             }
-            if ad.last_write.duration_since(now) > DERP_INACTIVE_CLEANUP_TIME {
-                to_close.push(*i);
+            let (os, or) = oneshot::channel();
+            match s.send(ActiveDerpMessage::GetLastWrite(os)).await {
+                Ok(_) => match or.await {
+                    Ok(last_write) => {
+                        if last_write.duration_since(now) > DERP_INACTIVE_CLEANUP_TIME {
+                            to_close.push(*i);
+                        }
+                    }
+                    Err(_) => {
+                        to_close.push(*i);
+                    }
+                },
+                Err(_) => {
+                    to_close.push(*i);
+                }
             }
         }
 
         let dirty = !to_close.is_empty();
-        debug!(
-            "closing {}/{} derps",
+        trace!(
+            "closing {} of {} derps",
             to_close.len(),
             self.active_derp.len()
         );
@@ -409,82 +629,34 @@ impl DerpActor {
     }
 
     async fn close_derp(&mut self, region_id: u16, why: &'static str) {
-        if let Some(ad) = self.active_derp.remove(&region_id) {
-            debug!(
-                "closing connection to derp-{} ({:?}), age {}s",
-                region_id,
-                why,
-                ad.create_time.elapsed().as_secs()
-            );
+        if let Some((s, t)) = self.active_derp.remove(&region_id) {
+            debug!(region_id, "closing connection: {}", why);
 
-            let ActiveDerp { c, cancel, .. } = ad;
-            c.close().await;
-            cancel.cancel();
+            s.send(ActiveDerpMessage::Shutdown).await.ok();
+            t.abort(); // ensure the task is shutdown
 
             inc!(MagicsockMetrics, num_derp_conns_removed);
         }
     }
 
     fn log_active_derp(&self) {
-        let now = Instant::now();
         debug!("{} active derp conns{}", self.active_derp.len(), {
             let mut s = String::new();
             if !self.active_derp.is_empty() {
                 s += ":";
-                for (node, ad) in self.active_derp_sorted() {
-                    s += &format!(
-                        " derp-{}=cr{},wr{}",
-                        node,
-                        now.duration_since(ad.create_time).as_secs(),
-                        now.duration_since(ad.last_write).as_secs()
-                    );
+                for node in self.active_derp_sorted() {
+                    s += &format!(" derp-{}", node,);
                 }
             }
             s
         });
     }
 
-    fn active_derp_sorted(&self) -> impl Iterator<Item = (u16, &'_ ActiveDerp)> + '_ {
+    fn active_derp_sorted(&self) -> impl Iterator<Item = u16> {
         let mut ids: Vec<_> = self.active_derp.keys().copied().collect();
         ids.sort();
 
         ids.into_iter()
-            .map(|id| (id, self.active_derp.get(&id).unwrap()))
-    }
-
-    /// Removes a DERP route entry previously added by add_derp_peer_route.
-    fn remove_derp_peer_routes(
-        &mut self,
-        peers: Vec<PublicKey>,
-        derp_id: u16,
-        dc: &derp::http::Client,
-    ) {
-        for peer in peers {
-            if let hash_map::Entry::Occupied(r) = self.derp_route.entry(peer) {
-                if r.get().derp_id == derp_id && &r.get().dc == dc {
-                    r.remove();
-                }
-            }
-        }
-    }
-
-    /// Adds DERP route entries, noting that peer was seen on DERP node `derp_id`, at least on the
-    /// connection identified by `dc`.
-    fn add_derp_peer_routes(
-        &mut self,
-        peers: Vec<PublicKey>,
-        derp_id: u16,
-        dc: derp::http::Client,
-    ) {
-        for peer in peers {
-            self.derp_route.insert(
-                peer,
-                DerpRoute {
-                    derp_id,
-                    dc: dc.clone(),
-                },
-            );
-        }
     }
 }
 
@@ -497,184 +669,10 @@ pub(super) struct DerpReadResult {
     pub(super) buf: Bytes,
 }
 
-/// Manages reading state for a single derp connection.
-#[derive(Debug)]
-struct ReaderState {
-    region: u16,
-    derp_client: derp::http::Client,
-    /// The set of senders we know are present on this connection, based on
-    /// messages we've received from the server.
-    peer_present: HashSet<PublicKey>,
-    backoff: backoff::exponential::ExponentialBackoff<backoff::SystemClock>,
-    last_packet_time: Option<Instant>,
-    last_packet_src: Option<PublicKey>,
-    cancel: CancellationToken,
-}
-
-#[derive(Debug)]
+#[derive(Debug, PartialEq, Eq)]
 pub(super) enum ReadResult {
-    Yield(DerpReadResult),
     Break,
     Continue,
-}
-
-#[derive(Debug)]
-pub(super) enum ReadAction {
-    None,
-    RemovePeerRoutes {
-        peers: Vec<PublicKey>,
-        region: u16,
-        derp_client: derp::http::Client,
-    },
-    AddPeerRoutes {
-        peers: Vec<PublicKey>,
-        region: u16,
-        derp_client: derp::http::Client,
-    },
-}
-
-impl ReaderState {
-    fn new(region: u16, cancel: CancellationToken, derp_client: derp::http::Client) -> Self {
-        ReaderState {
-            region,
-            derp_client,
-            cancel,
-            peer_present: HashSet::new(),
-            backoff: backoff::exponential::ExponentialBackoffBuilder::new()
-                .with_initial_interval(Duration::from_millis(10))
-                .with_max_interval(Duration::from_secs(5))
-                .build(),
-            last_packet_time: None,
-            last_packet_src: None,
-        }
-    }
-
-    async fn recv(&mut self) -> (ReadResult, ReadAction) {
-        let msg = tokio::select! {
-            msg = self.derp_client.recv_detail() => {
-                msg
-            }
-            _ = self.cancel.cancelled() => {
-                return (ReadResult::Break, ReadAction::None);
-            }
-        };
-        debug!(region_id=%self.region, ?msg, "derp.recv received");
-
-        match msg {
-            Err(err) => {
-                debug!(
-                    "[{:?}] derp.recv(derp-{}): {:?}",
-                    self.derp_client, self.region, err
-                );
-
-                // Forget that all these peers have routes.
-                let peers = self.peer_present.drain().collect();
-                let action = ReadAction::RemovePeerRoutes {
-                    peers,
-                    region: self.region,
-                    derp_client: self.derp_client.clone(),
-                };
-
-                if matches!(
-                    err,
-                    derp::http::ClientError::Closed | derp::http::ClientError::IPDisabled
-                ) {
-                    // drop client
-                    return (ReadResult::Break, action);
-                }
-
-                // If our DERP connection broke, it might be because our network
-                // conditions changed. Start that check.
-                // TODO:
-                // self.re_stun("derp-recv-error").await;
-
-                // Back off a bit before reconnecting.
-                match self.backoff.next_backoff() {
-                    Some(t) => {
-                        debug!("backoff sleep: {}ms", t.as_millis());
-                        time::sleep(t).await;
-                        (ReadResult::Continue, action)
-                    }
-                    None => (ReadResult::Break, action),
-                }
-            }
-            Ok((msg, conn_gen)) => {
-                // reset
-                self.backoff.reset();
-                let now = Instant::now();
-                if self.last_packet_time.is_none()
-                    || self.last_packet_time.as_ref().unwrap().elapsed() > Duration::from_secs(5)
-                {
-                    self.last_packet_time = Some(now);
-                }
-
-                match msg {
-                    derp::ReceivedMessage::ServerInfo { .. } => {
-                        info!("derp-{} connected; connGen={}", self.region, conn_gen);
-                        (ReadResult::Continue, ReadAction::None)
-                    }
-                    derp::ReceivedMessage::ReceivedPacket { source, data } => {
-                        trace!("[DERP] <- {} ({}b)", self.region, data.len());
-                        // If this is a new sender we hadn't seen before, remember it and
-                        // register a route for this peer.
-                        let action = if self.last_packet_src.is_none()
-                            || &source != self.last_packet_src.as_ref().unwrap()
-                        {
-                            // avoid map lookup w/ high throughput single peer
-                            self.last_packet_src = Some(source);
-                            let mut peers = Vec::new();
-                            if !self.peer_present.contains(&source) {
-                                self.peer_present.insert(source);
-                                peers.push(source);
-                            }
-                            ReadAction::AddPeerRoutes {
-                                peers,
-                                region: self.region,
-                                derp_client: self.derp_client.clone(),
-                            }
-                        } else {
-                            ReadAction::None
-                        };
-
-                        let res = DerpReadResult {
-                            region_id: self.region,
-                            src: source,
-                            buf: data,
-                        };
-                        (ReadResult::Yield(res), action)
-                    }
-                    derp::ReceivedMessage::Ping(data) => {
-                        // Best effort reply to the ping.
-                        let dc = self.derp_client.clone();
-                        let region = self.region;
-                        tokio::task::spawn(async move {
-                            if let Err(err) = dc.send_pong(data).await {
-                                info!("derp-{} send_pong error: {:?}", region, err);
-                            }
-                        });
-                        (ReadResult::Continue, ReadAction::None)
-                    }
-                    derp::ReceivedMessage::Health { .. } => {
-                        // health.SetDERPRegionHealth(regionID, m.Problem);
-                        (ReadResult::Continue, ReadAction::None)
-                    }
-                    derp::ReceivedMessage::PeerGone(key) => {
-                        let read_action = ReadAction::RemovePeerRoutes {
-                            peers: vec![key],
-                            region: self.region,
-                            derp_client: self.derp_client.clone(),
-                        };
-
-                        (ReadResult::Continue, read_action)
-                    }
-                    _ => {
-                        // Ignore.
-                        (ReadResult::Continue, ReadAction::None)
-                    }
-                }
-            }
-        }
-    }
 }
 
 /// Combines blobs into packets of at most MAX_PACKET_SIZE.
